Grouped PHP MySQL into one folder
Added a new project with a form login screen.  This form wil be used to
record and modify database information in MySQL.  Later I want to create
a second project with similar login form that will allow me to save all
my monthly bills to a database and then send by email to girlfriend on
her share of the expenses.